fixed GetByServiceID(). serviceID is an int, not an unsigned short, and was converted to 0 because the number was too large

+++ vdr/channels/ChannelManager.cpp
@@ -300,7 +300,7 @@ ChannelPtr cChannelManager::GetByNumber(int number, int skipGap /* = 0 */)
   return cChannel::EmptyChannel;
 }
 
-ChannelPtr cChannelManager::GetByServiceID(unsigned short serviceID, int source, int transponder)
+ChannelPtr cChannelManager::GetByServiceID(int serviceID, int source, int transponder)
 {
   assert(serviceID);
   assert(source);
